backend/requirements.txt
flask
flask-cors
flask-socketio
pymongo
python-dotenv

